http binary tests in place
