--- azure-functions-java-worker/src/it/java/com/microsoft/azure/webjobs/script/integrationTest/Function.java
@@ -1,40 +1,0 @@
-package com.microsoft.azure.webjobs.script.integrationTest;
-
-import java.util.*;
-import com.microsoft.azure.serverless.functions.annotation.*;
-import com.microsoft.azure.serverless.functions.*;
-
-/**
- * Azure Functions with HTTP Trigger.
- */
-public class Function {
-    /**
-     * This function listens at endpoint "/api/hello". Two ways to invoke it using "curl" command in bash:
-     * 1. curl -d "HTTP Body" {your host}/api/hello
-     * 2. curl {your host}/api/hello?name=HTTP%20Query
-     */
-    @FunctionName("hello")
-    public HttpResponseMessage<String> hello(
-            @HttpTrigger(name = "req", methods = {"get", "post"}, authLevel = AuthorizationLevel.ANONYMOUS) HttpRequestMessage<Optional<String>> request,
-            final ExecutionContext context) {
-        context.getLogger().info("Java HTTP trigger processed a request.");
-
-        // Parse query parameter
-        String query = request.getQueryParameters().get("name");
-        String name = request.getBody().orElse(query);
-
-        if (name == null) {
-            return request.createResponse(400, "Please pass a name on the query string or in the request body");
-        } else {
-            return request.createResponse(200, "Hello, " + name);
-        }
-    }
-
-    public static String testPrimitiveByteArray(byte[] data) {
-        return "Received " + data.length + " bytes byte[] data";
-    }
-
-    public static String testClassByteArray(Byte[] data) {
-        return "Received " + data.length + " bytes Byte[] data";
-    }
-}
