+++ azure-functions-java-worker/src/it/java/com/microsoft/azure/webjobs/script/test/functions/HttpBinary.java
@@ -1,0 +1,14 @@
+package com.microsoft.azure.webjobs.script.test.functions;
+
+
+public class HttpBinary {
+    public static String primitiveByteArray(byte[] data) {
+
+        return "Received " + data.length + " bytes byte[] data";
+    }
+
+    public static String classByteArray(Byte[] data) {
+
+        return "Received " + data.length + " bytes Byte[] data";
+    }
+}
