+++ azure-functions-java-worker/src/it/resources/classByteArray/function.json
@@ -1,6 +1,6 @@
 {
   "scriptFile": "../../azure-functions-java-worker-1.0.0-beta-4-tests.jar",
-  "entryPoint": "com.microsoft.azure.webjobs.script.integrationTest.Function.testClassByteArray",
+  "entryPoint": "com.microsoft.azure.webjobs.script.test.functions.HttpBinary.classByteArray",
   "bindings": [
     {
       "type": "httpTrigger",
